+++ pyteal/ast/frame.py
@@ -1,5 +1,5 @@
 from itertools import groupby
-from typing import TYPE_CHECKING, Optional
+from typing import TYPE_CHECKING, Optional, Final
 
 from pyteal.ast.expr import Expr
 from pyteal.ast.int import Int
@@ -11,6 +11,9 @@ from pyteal.ir import TealBlock, TealSimpleBlock, TealOp, Op
 
 if TYPE_CHECKING:
     from pyteal.compiler import CompileOptions
+
+
+MAX_FRAME_LOCAL_VARS: Final[int] = 128
 
 
 class LocalTypeSegment(Expr):
@@ -41,14 +44,7 @@ class LocalTypeSegment(Expr):
                 )
 
     def __teal__(self, options: "CompileOptions") -> tuple[TealBlock, TealSimpleBlock]:
-        if self.count == 1:
-            inst_srt, inst_end = self.auto_instance.__teal__(options)
-            return inst_srt, inst_end
-        else:
-            dupn_srt, dupn_end = DupN(self.auto_instance, self.count - 1).__teal__(
-                options
-            )
-            return dupn_srt, dupn_end
+        return DupN(self.auto_instance, self.count - 1).__teal__(options)
 
     def __str__(self) -> str:
         return f"(LocalTypeSegment: (type: {self.local_type}) (count: {self.count}))"
@@ -104,6 +100,14 @@ class ProtoStackLayout(Expr):
     def __str__(self) -> str:
         return f"(ProtoStackLayout: (args: {self.arg_stack_types}) (locals: {self.local_stack_types}))"
 
+    @classmethod
+    def from_proto(cls, proto: "Proto") -> "ProtoStackLayout":
+        return cls(
+            [TealType.anytype] * proto.num_args,
+            [TealType.anytype] * proto.num_returns,
+            proto.num_returns,
+        )
+
     def has_return(self) -> bool:
         return False
 
@@ -157,6 +161,9 @@ class Proto(Expr):
             raise TealInputError(
                 f"The number of returns provided to Proto must be >= 0 but {num_returns=}."
             )
+        self.num_args = num_args
+        self.num_returns = num_returns
+
         if mem_layout:
             if mem_layout.num_return_allocs > num_returns:
                 raise TealInternalError(
@@ -168,10 +175,10 @@ class Proto(Expr):
                     f"The number of arguments {num_args} should match with "
                     f"memory layout's number of arguments {len(mem_layout.arg_stack_types)}"
                 )
+        else:
+            mem_layout = ProtoStackLayout.from_proto(self)
 
-        self.num_args = num_args
-        self.num_returns = num_returns
-        self.mem_layout: Optional[ProtoStackLayout] = mem_layout
+        self.mem_layout: ProtoStackLayout = mem_layout
 
     def __teal__(self, options: "CompileOptions") -> tuple[TealBlock, TealSimpleBlock]:
         verifyProgramVersion(
@@ -181,8 +188,6 @@ class Proto(Expr):
         )
         op = TealOp(self, Op.proto, self.num_args, self.num_returns)
         proto_srt, proto_end = TealBlock.FromOp(options, op)
-        if not self.mem_layout:
-            return proto_srt, proto_end
         local_srt, local_end = self.mem_layout.__teal__(options)
         proto_end.setNextBlock(local_srt)
         return proto_srt, local_end
@@ -250,7 +255,7 @@ class FrameBury(Expr):
     ):
         super().__init__()
 
-        target_type = inferred_type if inferred_type else TealType.anytype
+        target_type = inferred_type if inferred_type is not None else TealType.anytype
         require_type(value, target_type)
 
         self.value = value
@@ -291,11 +296,7 @@ class FrameVar(AbstractVar):
         super().__init__()
         self.proto = under_proto
         self.frame_index = frame_index
-        self.stack_type = (
-            self.proto.mem_layout[frame_index]
-            if self.proto.mem_layout
-            else TealType.anytype
-        )
+        self.stack_type = self.proto.mem_layout[frame_index]
 
     def storage_type(self) -> TealType:
         return self.stack_type
@@ -322,6 +323,14 @@ class DupN(Expr):
     """
 
     def __init__(self, value: Expr, repetition: int):
+        """Create a DupN expression.
+
+        Args:
+            value: The value to be duplicated.
+            repetition: How many additional times the value should be added to the stack. At the end
+                of this operation, `repetition+1` elements will be added to the stack. Zero can be
+                specified here to indicate no duplication.
+        """
         super().__init__()
         require_type(value, TealType.anytype)
         if repetition < 0:
@@ -330,6 +339,15 @@ class DupN(Expr):
         self.repetition = repetition
 
     def __teal__(self, options: "CompileOptions") -> tuple[TealBlock, TealSimpleBlock]:
+        if self.repetition == 0:
+            # no duplication required
+            return self.value.__teal__(options)
+
+        if self.repetition == 1:
+            # use normal dup op for just 1 duplication
+            op = TealOp(self, Op.dup)
+            return TealBlock.FromOp(options, op, self.value)
+
         verifyProgramVersion(
             Op.dupn.min_version,
             options.version,